Start fixing precision losses issues: fix and merge fileOffsetToRVA and RVAToFileOffset into common functions to 32 and 64 bit targets (operating and returning 64 bit addresses and offsets)

+++ CRTFootPrints.h
@@ -16,7 +16,7 @@ typedef uint8_t AsmFootPrint[][FOOTPRINT_STRIDE];
 
 @interface MachOLayout (CRTFootPrints)
 
-- (bool) matchAsmAtOffset:(uint32_t)offset 
+- (bool) matchAsmAtOffset:(uint64_t)offset 
              asmFootPrint:(const AsmFootPrint)footprint 
                 lineCount:(NSUInteger)lineCount;
